+++ OmegaEngine/OEMaths/OEMaths_convert.cpp
@@ -6,6 +6,6 @@ namespace OEMaths
 	// some popular maths conversions (haven't decided were to locate these yet!)
 	float radians(const float deg)
 	{
-		return deg * M_PI / 180.0f;
+		return deg * (float)M_PI / 180.0f;
 	}
 }
